% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambdaMaxC}
\alias{lambdaMaxC}
\title{Calculates the maximum penalty coefficient lambda for which all 
coefficients become zero}
\description{
Calculates the maximum penalty coefficient lambda for which all 
coefficients become zero
}
