% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sparseSurvData.R
\docType{data}
\name{sparseSurvData}
\alias{sparseSurvData}
\title{Simulated sparse survival dataset}
\format{
An object of class data.frame including 100 observations (rows)
and 150 features (columns)
}
\usage{
data(sparseSurvData)
}
\description{
Simulated sparse survival dataset
}
\examples{
\donttest{
data(sparseSurvData)
}
}
\references{
Ralf Bender, Thomas Augustin, and Maria Blettner
(\href{https://onlinelibrary.wiley.com/doi/10.1002/sim.2059}{Generating survival times to simulate Cox proportional hazards models})
}
\keyword{datasets}
