% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-survData.R
\docType{data}
\name{survData}
\alias{survData}
\title{Simulated survival dataset}
\format{
An object of class data.frame including 1000 observations (rows)
and 10 features (columns)
}
\usage{
data(survData)
}
\description{
Simulated survival dataset
}
\examples{
\donttest{
data(survData)
}
}
\references{
Ralf Bender, Thomas Augustin, and Maria Blettner
(\href{https://onlinelibrary.wiley.com/doi/10.1002/sim.2059}{Generating survival times to simulate Cox proportional hazards models})
}
\keyword{datasets}
